requirements.txt
fastapi
uvicorn[standard]
pydantic
uuid
python-multipart
sqlalchemy
#passlib==1.7.4
#bcrypt==3.2.2
python-jose
python-multipart
psycopg[binary]